// src/app/interfaces/spec-data.ts
export interface SpecData {

    category: string;
    difficulty: string;
    question: string;
    answer_1: string;
    answer_2: string;
    answer_3: string;
    answer_4: string;
    correct_answer: string;
}
export interface ITrivia {
    question: string;
    a1: string;
    a2: string;
    a3: string;
    a4: string;
    correctA: string;
  }